pyproject.toml
[tool.poetry]
name = "aqi-project"
version = "0.1.0"
description = "AQI_project eco_sys"
authors = ["6510110513 <6510110513@psu.ac.th>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amqp = "5.2.0"
annotated-types = "0.7.0"
anyio = "4.6.0"
authlib = "1.3.2"
billiard = "4.2.1"
black = "23.12.1"
blinker = "1.8.2"
cachelib = "0.9.0"
celery = "5.4.0"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cryptography = "43.0.1"
dash = "2.18.1"
dash-bootstrap-components = "1.6.0"
dash-core-components = "2.0.0"
dash-html-components = "2.0.0"
dash-leaflet = "1.0.15"
dash-table = "5.0.0"
dill = "0.3.8"
diskcache = "5.6.3"
dlib = "19.24.6"
flask = "3.0.3"
flask-caching = "2.3.0"
flask-login = "0.6.3"
flask-sqlalchemy = "3.1.1"
flask-wtf = "1.2.1"
greenlet = "3.1.1"
h11 = "0.14.0"
httpcore = "0.17.3"
httpx = "0.24.1"
idna = "3.10"
importlib-metadata = "8.5.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
kombu = "5.4.2"
markdown = "3.7"
markupsafe = "2.1.5"
multiprocess = "0.70.16"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
numpy = "2.1.1"
opencv-python = "4.10.0.84"
packaging = "24.1"
pandas = "2.2.3"
pathspec = "0.12.1"
platformdirs = "4.3.6"
plotly = "5.24.1"
prompt-toolkit = "3.0.48"
psutil = "6.0.0"
psycopg2-binary = "2.9.9"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-settings = "2.5.2"
pydantic-core = "2.23.4"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
redis = "5.1.0"
requests = "2.32.3"
retrying = "1.3.4"
rq = "1.16.2"
setuptools = "75.1.0"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.35"
style = "1.1.0"
tenacity = "9.0.0"
typing-extensions = "4.12.2"
tzdata = "2024.2"
update = "0.0.1"
urllib3 = "2.2.3"
vine = "5.1.0"
wcwidth = "0.2.13"
werkzeug = "3.0.4"
wtforms = "3.1.2"
xlrd = "2.0.1"
zipp = "3.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
